fix #9760 feat(cirrus): Move demo app into pages section
